pyproject.toml
[tool.poetry]
name = "vector-search-test-task"
version = "0.1.0"
description = ""
authors = ["Adrian Dankiv <adr-007@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
weaviate-client = "^4.4.4"
funcy = "^2.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
